+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Fixed
+
+- Detect question-mark-only IAM action wildcards such as `sqs:?etQueueAttributes`
+
 ## [2.0.0] - 2026-08-23
 
 ### Fixed